refact: move formatMetricValue function into utils file
