+++ src/components/common/MetricBox.tsx
@@ -1,4 +1,9 @@
-import { cn, formatLastUpdatedLive, formatSnakeCaseToTitle } from "@/lib/utils";
+import {
+  cn,
+  formatLastUpdatedLive,
+  formatMetricValue,
+  formatSnakeCaseToTitle,
+} from "@/lib/utils";
 import React, { memo, useEffect, useRef, useState } from "react";
 import {
   Card,
@@ -13,19 +18,6 @@ import {
   valueDecreaseAnimation,
   valueIncreaseAnimation,
 } from "@/constants/animations";
-
-// Utility function to format metric values with units
-const formatMetricValue = (key: string, value: number): string => {
-  const unitMap: Record<string, string> = {
-    battery_level: "%",
-    temperature: " \u00B0C",
-    humidity: "%",
-    signal_strength: " dBm",
-  };
-
-  const unit = unitMap[key] || "";
-  return `${value.toFixed(2)}${unit}`;
-};
 
 interface MetricBoxProps {
   metricLabel: string;
